+++ CHANGELOG.md
@@ -10,6 +10,25 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [1.17.4]
+### Added
+- Added a flag `--fixed-param-names` to prevent certain parameters from being optimized during training.
+  This is useful if you want to keep pre-trained embeddings fixed during training.
+- Added a flag `--dry-run` to `sockeye.train` to not perform any actual training, but print statistics about the model
+  and mode of operation.
+
+## [1.17.3]
+### Changed
+- `sockeye.evaluate` can now handle multiple hypotheses files by simply specifying `--hypotheses file1 file2...`.
+For each metric the mean and standard deviation will be reported across files.
+
+## [1.17.2]
+### Added
+- Optionally store the beam search history to a `json` output using the `beam_store` output handler.
+
+### Changed
+- Use stack operator instead of expand_dims + concat in RNN decoder. Reduces memory usage.
+
 ## [1.17.1]
 ### Changed
  - Updated to [MXNet 1.1.0](https://github.com/apache/incubator-mxnet/tree/1.1.0)